Revert FirstTimeBitmapWriter changes to understand issue
